Move individual tests to test file
