+++ lefthk/src/config.rs
@@ -108,8 +108,8 @@ fn try_from(kb: Keybind, default_modifier: Vec<String>) -> Result<Vec<core_keybi
 
 #[derive(Debug, Serialize, Deserialize, PartialEq, Eq, Clone)]
 pub struct Config {
-    default_modifier: Vec<String>,
-    keybinds: Vec<Keybind>,
+    pub default_modifier: Vec<String>,
+    pub keybinds: Vec<Keybind>,
 }
 
 impl lefthk_core::config::Config for Config {
@@ -179,142 +179,5 @@ fn propagate_exit_chord(chords: Vec<&mut Keybind>, exit_chord: Option<Keybind>) 
                 .collect();
             propagate_exit_chord(sub_chords, parent_exit_chord);
         }
-    }
-}
-
-#[cfg(test)]
-mod test {
-    use lefthk_core::config::{Command, Config, Keybind};
-
-    use super::Config as Cfg;
-
-    #[test]
-    fn parse_config() {
-        let config = r#"#![enable(implicit_some)]
-Config(
-    default_modifier: ["Mod4", "Shift"],
-    keybinds: [
-        Keybind(
-            command: Execute("st -e htop"),
-            key: Key("x"),
-        ),
-        Keybind(
-            command: Execute("st -e btm"),
-            modifier: ["Mod4"],
-            key: Key("c"),
-        ),
-    ]
-)"#;
-        let conf = Cfg::from_string(config.to_string());
-        assert!(conf.is_ok());
-        let conf = conf.unwrap();
-        assert_eq!(conf.default_modifier.len(), 2);
-        assert_eq!(
-            conf.default_modifier,
-            vec!["Mod4".to_string(), "Shift".to_string()]
-        );
-        let conf_mapped = conf.mapped_bindings();
-
-        // Verify default modifier implementation
-        let default_keybind = conf_mapped.first().unwrap();
-        assert_eq!(default_keybind.modifier.len(), 2);
-        assert_eq!(default_keybind.modifier, conf.default_modifier);
-
-        // Verify own implementation
-        let custom_keybind = conf_mapped.last().unwrap();
-        assert_eq!(custom_keybind.modifier.len(), 1);
-        assert_eq!(custom_keybind.modifier, vec!["Mod4".to_string()]);
-    }
-
-    #[test]
-    fn parse_empty_config() {
-        let config = r#"Config(
-    default_modifier: ["Mod4", "Shift"],
-    keybinds: []
-)"#;
-        let conf = Cfg::from_string(config.to_string());
-        assert!(conf.is_ok());
-        let conf = conf.unwrap();
-        assert_eq!(conf.default_modifier.len(), 2);
-        assert_eq!(
-            conf.default_modifier,
-            vec!["Mod4".to_string(), "Shift".to_string()]
-        );
-        let conf_mapped = conf.mapped_bindings();
-
-        // Verify implementation
-        assert_eq!(conf_mapped.len(), 0);
-    }
-
-    #[test]
-    fn parse_none_config() {
-        // Define empty string
-        let conf = Cfg::from_string(String::new());
-        assert!(conf.is_err());
-    }
-
-    #[test]
-    fn parse_sub_keybind_config() {
-        let config = r#"#![enable(implicit_some)]
-Config(
-    default_modifier: ["Mod4", "Shift"],
-    keybinds: [
-        Keybind(
-            command: Chord([
-                Keybind(
-                    command: Execute("st -e htop"),
-                    modifier: ["Mod4"],
-                    key: Key("c"),
-                ),
-            ]),
-            modifier: ["Mod4"],
-            key: Key("c"),
-        ),
-        Keybind(
-            command: Chord([
-                Keybind(
-                    command: Execute("st -e htop"),
-                    key: Key("c"),
-                ),
-            ]),
-            key: Key("c"),
-        ),
-    ]
-)"#;
-        let conf = Cfg::from_string(config.to_string());
-        assert!(conf.is_ok());
-        let conf = conf.unwrap();
-        assert_eq!(conf.default_modifier.len(), 2);
-        assert_eq!(
-            conf.default_modifier,
-            vec!["Mod4".to_string(), "Shift".to_string()]
-        );
-        let conf_mapped = conf.mapped_bindings();
-
-        // Verify default modifier implementation
-        let default_keybind = conf_mapped.last().unwrap();
-        assert_eq!(default_keybind.modifier.len(), 2);
-        assert_eq!(default_keybind.modifier, conf.default_modifier);
-        assert_eq!(
-            default_keybind.command,
-            Command::Chord(vec![Keybind {
-                command: Command::Execute("st -e htop".to_string()),
-                modifier: vec!["Mod4".to_string(), "Shift".to_string()],
-                key: "c".to_string(),
-            }])
-        );
-
-        // Verify custom modifier implementation
-        let custom_keybind = conf_mapped.first().unwrap();
-        assert_eq!(custom_keybind.modifier.len(), 1);
-        assert_eq!(custom_keybind.modifier, vec!["Mod4".to_string()]);
-        assert_eq!(
-            custom_keybind.command,
-            Command::Chord(vec![Keybind {
-                command: Command::Execute("st -e htop".to_string()),
-                modifier: vec!["Mod4".to_string()],
-                key: "c".to_string(),
-            }])
-        );
     }
 }
